# api/swagger.yaml
openapi: 3.0.3
info:
  title: Budget Planner API
  description: A comprehensive budget planning and expense tracking API
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/personal/webapp
    email: support@budgetplanner.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8082
    description: Development server
  - url: http://instance-agent.subnet05071228.vcn05071228.oraclevcn.com:8082
    description: Production server
...
paths:
...
  /api/v1/summary:
    get:
      tags:
        - Legacy
      summary: Get monthly summary
      description: Get financial summary for specified month
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2025
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 8
      responses:
        '200':
          description: Monthly summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Summary'
        '400':
          description: Invalid year or month
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/monthly-data:
    get:
      tags:
        - Monthly Data
      summary: Get comprehensive monthly data
      description: Get all financial data for specified month including income sources, budget sources, and expenses
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2025
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 8
      responses:
        '200':
          description: Comprehensive monthly data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyData'
        '400':
          description: Invalid year or month
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/income-sources:
    get:
      tags:
        - Income Sources
      summary: List income sources
      description: Get all income sources for specified month
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2025
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 8
      responses:
        '200':
          description: List of income sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IncomeSource'
        '400':
          description: Invalid year or month
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Income Sources
      summary: Create income source
      description: Create a new income source for specified month
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncomeSourceRequest'
      responses:
        '201':
          description: Income source created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncomeSource'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/income-sources/{id}:
    put:
      tags:
        - Income Sources
      summary: Update income source
      description: Update an existing income source
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSourceRequest'
      responses:
        '200':
          description: Income source updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          description: Invalid ID or request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Income Sources
      summary: Delete income source
      description: Delete an existing income source
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Income source deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/budget-sources:
    get:
      tags:
        - Budget Sources
      summary: List budget sources
      description: Get all budget sources for specified month
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2025
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 8
      responses:
        '200':
          description: List of budget sources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BudgetSource'
        '400':
          description: Invalid year or month
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Budget Sources
      summary: Create budget source
      description: Create a new budget source for specified month
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBudgetSourceRequest'
      responses:
        '201':
          description: Budget source created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetSource'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/budget-sources/{id}:
    put:
      tags:
        - Budget Sources
      summary: Update budget source
      description: Update an existing budget source
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSourceRequest'
      responses:
        '200':
          description: Budget source updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          description: Invalid ID or request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Budget Sources
      summary: Delete budget source
      description: Delete an existing budget source
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Budget source deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/expenses:
    get:
      tags:
        - Legacy
      summary: List expenses
      description: Get all expenses for specified month
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2025
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 8
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
        '400':
          description: Invalid year or month
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Legacy
      summary: Create expense
      description: Create a new expense
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 1
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/expenses/{id}:
    delete:
      tags:
        - Legacy
      summary: Delete expense
      description: Delete an existing expense
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Expense deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/salary:
    post:
      tags:
        - Legacy
      summary: Set monthly salary
      description: Set or update salary for specified month
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - year
                - month
                - amount_cents
              properties:
                year:
                  type: integer
                  example: 2025
                month:
                  type: integer
                  minimum: 1
                  maximum: 12
                  example: 8
                amount_cents:
                  type: integer
                  format: int64
                  example: 500000
                  description: Amount in cents (e.g., 500000 = $5000.00)
      responses:
        '200':
          description: Salary updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/budget:
    post:
      tags:
        - Legacy
      summary: Set monthly budget
      description: Set or update budget for specified month
      security:
        - APIKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - year
                - month
                - amount_cents
              properties:
                year:
                  type: integer
                  example: 2025
                month:
                  type: integer
                  minimum: 1
                  maximum: 12
                  example: 8
                amount_cents:
                  type: integer
                  format: int64
                  example: 400000
                  description: Amount in cents (e.g., 400000 = $4000.00)
      responses:
        '200':
          description: Budget updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    APIKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session_id
      description: Session-based authentication
  schemas:
...
    IncomeSource:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Primary Job"
        year:
          type: integer
          example: 2025
        month:
          type: integer
          minimum: 1
          maximum: 12
          example: 8
        amount_cents:
          type: integer
          format: int64
          example: 500000
          description: Amount in cents (e.g., 500000 = $5000.00)
        created_at:
          type: string
          format: date-time
          example: "2025-08-08T21:56:13Z"
        updated_at:
          type: string
          format: date-time
          example: "2025-08-08T21:57:54Z"
...
    Summary:
      type: object
      properties:
        year:
          type: integer
          example: 2025
        month:
          type: integer
          minimum: 1
          maximum: 12
          example: 8
        salary_cents:
          type: integer
          format: int64
          example: 500000
          description: Total salary in cents
        budget_cents:
          type: integer
          format: int64
          example: 400000
          description: Total budget in cents
        expense_cents:
          type: integer
          format: int64
          example: 350000
          description: Total expenses in cents
        remaining_cents:
          type: integer
          format: int64
          example: 150000
          description: Remaining amount in cents
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"